+++ CovidProofApp/src/main/java/com/covidproof/service/AadharcardService.java
@@ -4,6 +4,6 @@ import com.covidproof.model.Entity.AadharCard;
 
 public interface AadharcardService {
 	public AadharCard addAadharCard(AadharCard ac);
-	public AadharCard getAadharCard(Integer num);
-	public Boolean checkIfRegistered(Integer num);
+	public AadharCard getAadharCard(Long num);
+	public Boolean checkIfRegistered(Long num);
 }
